Plotting for each task
Update configurations for pretraining in cluster environment

+++ eyemind/dataloading/load_dataset.py
@@ -124,7 +124,7 @@ def get_samplers():
 
 def binarize_labels(y_data, threshold=0.5):
     # generate 0 if below threshold, 1 if above
-    y_data = (y_data > threshold).int()
+    y_data = (y_data > threshold)
     return y_data
 
 # def stratified_group_split(X, y, folds=4):
